using the memory backend instead, splitted settings into formats and settings, clarified usage of the two settings objects
