+++ data/org.soundconverter.gschema.xml
@@ -126,8 +126,8 @@
       <summary></summary>
       <description></description>
     </key>
-    <key name="limit-jobs" type="i">
-      <default>0</default>
+    <key name="limit-jobs" type="b">
+      <default>false</default>
       <summary>Limit jobs</summary>
       <description></description>
     </key>
