+++ tests/unittests.py
@@ -19,7 +19,7 @@ from gi.repository import Gst, Gio, Gtk, GLib
 Gst.init([None] + [a for a in sys.argv[1:] if '-gst' in a])
 
 import inspect
-from soundconverter.settings import settings
+from soundconverter.settings import settings, set_gio_settings, get_gio_settings
 from soundconverter.namegenerator import TargetNameGenerator
 from soundconverter.soundfile import SoundFile
 from soundconverter.fileoperations import filename_to_uri, unquote_filename, beautify_uri
@@ -31,12 +31,13 @@ from importlib.machinery import SourceFileLoader
 
 DEFAULT_SETTINGS = settings.copy()
 
+backend = Gio.memory_settings_backend_new()
+gio_settings = Gio.Settings.new_with_backend('org.soundconverter', backend)
+set_gio_settings(gio_settings)
+
 # tests will control gtk main iterations
 Gtk.main = gtk_iteration
 Gtk.main_quit = lambda: None
-
-gio_settings = Gio.Settings(schema='org.soundconverter')
-gio_settings_original_mapping = {key: gio_settings.get_value(key) for key in gio_settings.keys()}
 
 
 def launch(argv=[]):
@@ -185,14 +186,6 @@ class GUI(unittest.TestCase):
             shutil.rmtree("tests/tmp")
         os.makedirs("tests/tmp", exist_ok=True)
 
-    @classmethod
-    def tearDownClass(cls):
-        """Restore the settings that can be configured over the UI to the users specified values."""
-        for key in gio_settings_original_mapping:
-            gio_settings.set_value(key, gio_settings_original_mapping[key])
-        # apparently set_value needs some time to get applied
-        time.sleep(0.1)
-
     def tearDown(self):
         win[0].close()
         reset_settings()
@@ -211,12 +204,12 @@ class GUI(unittest.TestCase):
 
         # setup for conversion
         window.prefs.change_mime_type('audio/ogg; codecs=opus')
-        window.prefs.settings.set_boolean('create-subfolders', False)
-        window.prefs.settings.set_boolean('same-folder-as-input', False)
-        window.prefs.settings.set_string('selected-folder', os.path.abspath("tests/tmp"))
-        window.prefs.settings.set_int('name-pattern-index', 0)
-        window.prefs.settings.set_boolean('replace-messy-chars', True)
-        window.prefs.settings.set_boolean('delete-original', False)
+        get_gio_settings().set_boolean('create-subfolders', False)
+        get_gio_settings().set_boolean('same-folder-as-input', False)
+        get_gio_settings().set_string('selected-folder', os.path.abspath("tests/tmp"))
+        get_gio_settings().set_int('name-pattern-index', 0)
+        get_gio_settings().set_boolean('replace-messy-chars', True)
+        get_gio_settings().set_boolean('delete-original', False)
 
         # start conversion
         window.on_convert_button_clicked()
